+++ app/src/main/java/kr/co/connect/boostcamp/livewhere/ui/map/interfaces/OnMapHistoryListener.kt
@@ -1,4 +1,4 @@
-package kr.co.connect.boostcamp.livewhere.ui.map
+package kr.co.connect.boostcamp.livewhere.ui.map.interfaces
 
 import com.naver.maps.map.overlay.CircleOverlay
 import com.naver.maps.map.overlay.InfoWindow
