+++ app/src/main/java/kr/co/connect/boostcamp/livewhere/ui/map/interfaces/OnSearchTrigger.kt
@@ -1,4 +1,4 @@
-package kr.co.connect.boostcamp.livewhere.ui.map
+package kr.co.connect.boostcamp.livewhere.ui.map.interfaces
 
 import android.view.View
