+++ app/src/main/java/kr/co/connect/boostcamp/livewhere/ui/map/view/PlaceImageViewMotionLayout.kt
@@ -1,4 +1,5 @@
-package kr.co.connect.boostcamp.livewhere.ui.map
+package kr.co.connect.boostcamp.livewhere.ui.map.view
+
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -7,29 +8,25 @@ import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
 import androidx.constraintlayout.motion.widget.MotionLayout
-import kotlinx.android.synthetic.main.activity_map.view.*
+import kr.co.connect.boostcamp.livewhere.R
 
-class BackdropMotionLayout(context: Context, attributeSet: AttributeSet? = null) : MotionLayout(context, attributeSet) {
+class PlaceImageViewMotionLayout (context: Context, attributeSet: AttributeSet? = null) : MotionLayout(context, attributeSet) {
+    private val viewToDetectTouch by lazy {
+        findViewById<View>(R.id.iv_place_image)
+    }
+
     private val viewRect = Rect()
     private var touchStarted = false
-
     init {
         setTransitionListener(object : MotionLayout.TransitionListener {
-            override fun onTransitionTrigger(p0: MotionLayout?, p1: Int, p2: Boolean, p3: Float) {
+            override fun onTransitionTrigger(p0: MotionLayout?, p1: Int, p2: Boolean, p3: Float) {}
 
-            }
+            override fun onTransitionStarted(p0: MotionLayout?, p1: Int, p2: Int) {}
 
-            override fun onTransitionStarted(p0: MotionLayout?, p1: Int, p2: Int) {
+            override fun onTransitionChange(p0: MotionLayout?, p1: Int, p2: Int, progress: Float) {}
 
-            }
-
-            override fun onTransitionChange(p0: MotionLayout?, p1: Int, p2: Int, p3: Float) {
-
-            }
-
-            override fun onTransitionCompleted(motionLayout: MotionLayout?, currentId: Int) {
+            override fun onTransitionCompleted(motionLayout: MotionLayout?, motionId: Int) {
                 touchStarted = false
-
             }
         })
     }
@@ -43,11 +40,8 @@ class BackdropMotionLayout(context: Context, attributeSet: AttributeSet? = null)
             }
         }
         if (!touchStarted) {
-            ll_place_detail.getHitRect(viewRect)
+            viewToDetectTouch.getHitRect(viewRect)
             touchStarted = viewRect.contains(event.x.toInt(), event.y.toInt())
-            if(touchStarted){
-                ml_place_image.iv_place_image.visibility = View.GONE
-            }
         }
         return touchStarted && super.onTouchEvent(event)
     }
